move perpetual calculation functions to perpetuals/lib package
